CDD-2230: Moved components into the view folder as a way to logically group them
