+++ src/app/(pages)/switchboard/(pages)/weather-health-alerts/page.tsx
@@ -1,8 +1,8 @@
 import { cookies } from 'next/headers'
 
 import { View } from '@/app/components/ui/ukhsa'
-import { BackLink } from '@/app/components/ui/ukhsa/BackLink/Backlink'
-import { Heading } from '@/app/components/ui/ukhsa/Heading/Heading'
+import { BackLink } from '@/app/components/ui/ukhsa/View/BackLink/Backlink'
+import { Heading } from '@/app/components/ui/ukhsa/View/Heading/Heading'
 import { UKHSA_SWITCHBOARD_COOKIE_NAME } from '@/app/constants/app.constants'
 
 import { StatusSelect } from '../../components/StatusSelect'
