+++ src/app/components/cms/pages/Feedback.tsx
@@ -1,7 +1,7 @@
 import { getFeedbackPage } from '@/app/utils/cms'
 
 import { View } from '../../ui/ukhsa'
-import { Heading } from '../../ui/ukhsa/Heading/Heading'
+import { Heading } from '../../ui/ukhsa/View/Heading/Heading'
 import Feedback from '../Feedback/Feedback'
 
 export default async function FeedbackPage() {
