+++ src/app/components/cms/pages/Topic.tsx
@@ -8,10 +8,10 @@ import { PageComponentBaseProps } from '@/app/types'
 import { renderCard } from '@/app/utils/cms.utils'
 import { clsx } from '@/lib/clsx'
 
-import { Description } from '../../ui/ukhsa/Description/Description'
-import { Heading } from '../../ui/ukhsa/Heading/Heading'
-import { LastUpdated } from '../../ui/ukhsa/LastUpdated/LastUpdated'
 import { RelatedLinksWrapper } from '../../ui/ukhsa/RelatedLinks/RelatedLinksWrapper'
+import { Description } from '../../ui/ukhsa/View/Description/Description'
+import { Heading } from '../../ui/ukhsa/View/Heading/Heading'
+import { LastUpdated } from '../../ui/ukhsa/View/LastUpdated/LastUpdated'
 
 export default async function TopicPage({
   slug,
